# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OfficeQueueManagement API",
    "version": "1.0.0",
    "description": "OpenAPI spec for the Office Queue Management server"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local server" }
  ],
  "paths": {
    "/api/services": {
      "get": {
        "summary": "Get available services",
        "responses": {
          "200": {
            "description": "A list of services",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "services": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Service" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets": {
      "post": {
        "summary": "Create a ticket for a service",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["serviceId"],
                "properties": {
                  "serviceId": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ticketId": { "type": "integer" },
                    "serviceId": { "type": "integer" },
                    "serviceType": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad request" },
          "500": { "description": "Internal server error" }
        }
      }
    }
  ,
    "/api/counters": {
      "get": {
        "summary": "Get list of counters",
        "responses": {
          "200": {
            "description": "Array of counter ids",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "counters": {
                      "type": "array",
                      "items": { "type": "integer" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/counters/{id}/tickets": {
      "post": {
        "summary": "Serve next customer at counter",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "201": {
            "description": "Ticket assigned to counter",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "ticketId": { "type": "integer", "nullable": true } } }
              }
            }
          },
          "404": { "description": "Counter not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/counters/{id}/queues": {
      "get": {
        "summary": "Get queues for each service this counter handles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Queues per service type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "queues": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "integer" },
                            "issuetime": { "type": "string" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Counter not found" },
          "500": { "description": "Internal server error" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Service": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "type": { "type": "string" }
        }
      },
      "TicketResponse": {
        "type": "object",
        "properties": {
          "ticketId": { "type": "integer" }
        }
      }
    }
  }
}